Update to incorporate added argument for genelist file in subscript3.2
The genelist file is now provided as the only argument following the command to run this wrapper script. The user can set the genelist by either specifying the name of the txt file in the directory containing the 1_New_References folder, or just using the string "Angiosperms353"; which automatically creates and uses an Angiosperms353 txtfile without the user having to create it themselves.